split api (and tests) into routes/subfiles
